temporary version
use summary data of 'the' and 'pi' (NOT 'gam'). Then a Newton method is used to optimize gim-type likelihood subject to (bet, alp, the, lam). No inner loop and I did not resolve the dual problem. The Newton module converged well and fast even for 20 SNPs. However, most of the time it converged to a saddle point with an indefinite hessian matrix. I push this version to github for archive purpose. I will try to resolve dual problem in next version.